requirements.txt
transformers>=4.46.0
trl>=0.11.0
peft>=0.13.0
accelerate>=1.0.0
deepspeed>=0.15.0
datasets>=2.20.0
evaluate>=0.4.3
pandas>=2.2.0
pyyaml>=6.0.2
pytest>=8.3.0
